+++ isodata/base.py
@@ -67,6 +67,12 @@ class ISOBase:
     def get_historical_supply(self, date):
         raise NotImplementedError()
 
+    def get_battery_today(self):
+        raise NotImplementedError()
+
+    def get_historical_battery(self, date):
+        raise NotImplementedError()
+
     def _latest_lmp_from_today(self, market, locations):
         lmp_df = self.get_lmp_today(market, locations)
         # Assume sorted in ascending order