+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Added load forecasting to NYISO, PJM, CAISO, ISONE, Ercot, SPP, MISO
+- Add battery charging and discharging to CAISO
 - Removed yesterday methods. Use `get_historical_*()` instead
 
 ## v0.6.0 - Aug 17, 2022